Upstream is 37 commits ahead, and the move conflicts in one place: the
cast_expr branch of the C backend. That branch now reads its argument as
g.a.nodes[int(cast_arg_id)] instead of going through g.a.child_node, so
the 128-bit cast hook sits directly under the new binding and passes the
child id the way its neighbours do.

+++ vlib/builtin/backtraces_windows.c.v
@@ -114,10 +114,10 @@ fn print_backtrace_skipping_top_frames_msvc(skipframes int) bool {
 		}
 		for i in 0 .. frames {
 			frame_addr := backtraces[i]
-			if C.SymFromAddr(handle, frame_addr, &offset, si) == 1 {
+			if C.SymFromAddr(handle, u64(frame_addr), &offset, si) == 1 {
 				nframe := frames - i - 1
 				mut lineinfo := ''
-				if C.SymGetLineFromAddr64(handle, frame_addr, &offset, &sline64) == 1 {
+				if C.SymGetLineFromAddr64(handle, u64(frame_addr), &offset, &sline64) == 1 {
 					file_name := unsafe { tos3(sline64.f_file_name) }
 					lnumber := sline64.f_line_number
 					lineinfo = file_name + ':' + i64(lnumber).str()
